requirements.txt
# Core dependencies
networkx>=3.2

# LLM Providers (install at least one)
anthropic>=0.40.0          # For Claude (default)
openai>=1.0.0            # For OpenAI / OpenAI-compatible endpoints
requests>=2.31.0         # For Ollama (local models)

# Optional
python-dotenv>=1.0.0     # For .env file support

# File Processing - 文件处理依赖
pypdf>=5.0.0             # PDF 文本提取
python-docx>=1.0.0      # Word DOCX 处理
openpyxl>=3.1.0          # Excel XLSX 处理
python-pptx>=1.0.0      # PowerPoint PPTX 处理
beautifulsoup4>=4.12.0   # HTML 解析
lxml>=5.0.0              # HTML 和 XML 解析
trafilatura>=1.6.0       # 高级网页内容提取
pillow>=10.0.0           # 图片处理
paddleocr>=2.7.0        # 图片 OCR (中文支持)

# Web Application (Demo)
flask>=3.0.0
flask-cors>=4.0.0
